+++ examples/plot_boxplot.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 import urllib.request
-
-from sqlalchemy import create_engine
+from sql.connection import Connection
 
 from sql import plot
 
@@ -12,6 +11,6 @@ if not Path("iris.csv").is_file():
         "iris.csv",
     )
 
-conn = create_engine("duckdb://")
+conn = Connection.from_connect_str("duckdb://").session
 
 plot.boxplot("iris.csv", "petal width", conn=conn)