+++ src/tests/integration/test_generic_db_operations.py
@@ -40,7 +40,10 @@ def mock_log_api(monkeypatch):
         ("ip_with_mySQL", 3),
         ("ip_with_mariaDB", 3),
         ("ip_with_SQLite", 3),
-        ("ip_with_duckDB", 3),
+        pytest.param(
+            "ip_with_duckDB",
+            3,
+        ),
     ],
 )
 def test_query_count(ip_with_dynamic_db, excepted, request):
@@ -67,7 +70,10 @@ def test_query_count(ip_with_dynamic_db, excepted, request):
         ("ip_with_mySQL", 15),
         ("ip_with_mariaDB", 15),
         ("ip_with_SQLite", 15),
-        ("ip_with_duckDB", 15),
+        pytest.param(
+            "ip_with_duckDB",
+            15,
+        ),
     ],
 )
 def test_create_table_with_indexed_df(ip_with_dynamic_db, excepted, request):