+++ src/tests/test_inspect.py
@@ -1,5 +1,4 @@
 from inspect import getsource
-import sqlite3
 import pytest
 from functools import partial
 
@@ -7,18 +6,17 @@ from sql import inspect, connection
 
 
 @pytest.fixture
-def sample_db(tmp_empty):
-    conn = connection.Connection.from_connect_str("sqlite://")
+def sample_db(ip):
+    ip.run_cell("%sql sqlite://")
+    ip.run_cell("%sql CREATE TABLE one (x INT, y TEXT)")
+    ip.run_cell("%sql CREATE TABLE another (i INT, j TEXT)")
+    ip.run_cell("%sql sqlite:///my.db")
+    ip.run_cell("%sql CREATE TABLE uno (x INT, y TEXT)")
+    ip.run_cell("%sql CREATE TABLE dos (i INT, j TEXT)")
+    ip.run_cell("%sql --close sqlite:///my.db")
+    ip.run_cell("%sql sqlite://")
 
-    conn.session.execute("CREATE TABLE one (x INT, y TEXT)")
-    conn.session.execute("CREATE TABLE another (i INT, j TEXT)")
-
-    conn_mydb = sqlite3.connect("my.db")
-    conn_mydb.execute("CREATE TABLE uno (x INT, y TEXT)")
-    conn_mydb.execute("CREATE TABLE dos (i INT, j TEXT)")
-    conn_mydb.close()
-
-    conn.session.execute("ATTACH DATABASE 'my.db' AS schema")
+    ip.run_cell("%sql ATTACH DATABASE 'my.db' AS schema")
 
 
 @pytest.mark.parametrize(
@@ -92,10 +90,9 @@ def test_get_column(sample_db, name, first, second, schema):
         ],
     ],
 )
-def test_nonexistent_table(name, schema, error):
+def test_nonexistent_table(sample_db, name, schema, error):
     with pytest.raises(ValueError) as excinfo:
         inspect.get_columns(name, schema)
-
     assert error.lower() in str(excinfo.value).lower()