+++ src/tests/test_magic.py
@@ -19,7 +19,7 @@ from conftest import runsql
 
 def test_memory_db(ip):
     assert runsql(ip, "SELECT * FROM test;")[0][0] == 1
-    assert runsql(ip, "SELECT * FROM test;")[1]["name"] == "bar"
+    assert runsql(ip, "SELECT * FROM test;")[1][1] == "bar"
 
 
 def test_html(ip):