+++ src/tests/test_telemetry.py
@@ -2,10 +2,9 @@ from pathlib import Path
 from unittest.mock import ANY, Mock
 import pytest
 import urllib.request
-from sqlalchemy import create_engine
 from sql.telemetry import telemetry
 from sql import plot
-
+from sql.connection import Connection
 
 # Ref: https://pytest.org/en/7.2.x/how-to/tmp_path.html#
 # Utilize tmp directory to store downloaded csv
@@ -40,7 +39,7 @@ def simple_file_path_penguins(tmpdir):
 
 @pytest.fixture
 def simple_db_conn():
-    return create_engine("duckdb://")
+    return Connection.from_connect_str("duckdb://").session
 
 
 @pytest.fixture